requirements.txt
celery==5.2.7
Django==3.1.7
djangorestframework==3.12.4
jsonschema==4.5.1
Pillow==8.4.0
psycopg2==2.9.1
redis==4.3.3